make utlils folder, move firepads.js to utils
